+++ src/Components/Web/src/Forms/InputChoice.cs
@@ -1,0 +1,46 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System;
+using System.Diagnostics.CodeAnalysis;
+using System.Globalization;
+
+namespace Microsoft.AspNetCore.Components.Forms
+{
+    /// <summary>
+    /// Serves as a base for inputs that have a group of selectable choices.
+    /// </summary>
+    public abstract class InputChoice<TValue> : InputBase<TValue>
+    {
+        private static readonly Type? _nullableUnderlyingType = Nullable.GetUnderlyingType(typeof(TValue));
+
+        /// <inheritdoc />
+        protected override bool TryParseValueFromString(string? value, [MaybeNull] out TValue result, [NotNullWhen(false)] out string? validationErrorMessage)
+        {
+            if (typeof(TValue) == typeof(string))
+            {
+                result = (TValue)(object?)value;
+                validationErrorMessage = null;
+                return true;
+            }
+            else if (typeof(TValue).IsEnum || (_nullableUnderlyingType != null && _nullableUnderlyingType.IsEnum))
+            {
+                var success = BindConverter.TryConvertTo<TValue>(value, CultureInfo.CurrentCulture, out var parsedValue);
+                if (success)
+                {
+                    result = parsedValue;
+                    validationErrorMessage = null;
+                    return true;
+                }
+                else
+                {
+                    result = default;
+                    validationErrorMessage = $"The {FieldIdentifier.FieldName} field is not valid.";
+                    return false;
+                }
+            }
+
+            throw new InvalidOperationException($"{GetType()} does not support the type '{typeof(TValue)}'.");
+        }
+    }
+}
